Save the private key in an output parameter instead of a file
https://github.com/wizeline/project-lab/pull/329\#discussion_r882106254
I suggest saving the private key in an output parameter instead of a file.
https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions\#setting-an-output-parameter

NOTE: I changed the command because I realized it could be shortened even more by using the -r flag in jq to get rid for the need of using the tr and sed commands. The head -n -1 will discard the last line output by jq, which is just an empty line and is not needed

That way, you can reference it in the key and ssh_private_key parameter for the SSH and SFTP actions like this:

key: ${{ steps.get-lightsail-ssh-key.outputs.SSH_PRIVATE_KEY }}